+++ src/index.ts
@@ -4,36 +4,45 @@ import * as cache from '@actions/cache'
 import path from 'node:path'
 import { Chalk } from 'chalk'
 
-import { getInstallSettings, runCargoInstall } from './install'
+import { type ResolvedVersion, getInstallSettings, runCargoInstall } from './install'
 import { parseInput } from './parse'
-import { resolveVersion } from './resolve'
+import { resolveRegistryVersion } from './resolve/registry'
+import { resolveGitCommit } from './resolve/git'
 
 const chalk = new Chalk({ level: 3 })
 
 async function run (): Promise<void> {
   const input = parseInput()
 
-  core.startGroup(chalk.bold(`Installing ${input.crate} ...`))
-  core.info('Fetching crate information on crates.io ...')
-  const version = await resolveVersion(input)
-  const install = getInstallSettings(input, version)
+  // Resolve crate version and try to restore from cache
+  core.startGroup(chalk.bold(`Installing ${input.crate}...`))
+  const version: ResolvedVersion = input.source.type === 'registry'
+    ? await resolveRegistryVersion(input.crate, input.source.version)
+    : await resolveGitCommit(input.source)
 
+  const install = getInstallSettings(input, version)
   core.info('Installation settings:')
-  core.info(`   version: ${version}`)
+  if ('version' in version) {
+    core.info(`   version: ${version.version}`)
+  } else {
+    core.info(`   repository: ${version.repository}`)
+    core.info(`   commit: ${version.commit}`)
+  }
   core.info(`   path: ${install.path}`)
   core.info(`   key: ${install.cacheKey}`)
 
   await io.mkdirP(install.path)
   const restored = await cache.restoreCache([install.path], install.cacheKey)
-
   core.endGroup()
 
+  // Check if the crate has been restored from cache
   let cacheHit = false
   if (restored !== undefined) {
     core.info(`Restored ${input.crate} from cache.`)
     cacheHit = true
   } else {
-    core.startGroup(`No cached version found, installing ${input.crate} using cargo ...`)
+    // Install the crate if it wasn't restored from cache
+    core.startGroup(`No cached version found, installing ${input.crate} using cargo...`)
     await runCargoInstall(input, version, install)
 
     try {
@@ -45,11 +54,17 @@ async function run (): Promise<void> {
     core.endGroup()
   }
 
+  // Add the crate's binary directory to PATH
   core.addPath(path.join(install.path, 'bin'))
   core.info(`Added ${install.path}/bin to PATH.`)
-  core.info(chalk.green(`Installed ${input.crate} ${version}.`))
 
-  core.setOutput('version', version)
+  if ('version' in version) {
+    core.info(chalk.green(`Installed ${input.crate} ${version.version}.`))
+  } else {
+    core.info(chalk.green(`Installed ${input.crate} from ${version.repository} at ${version.commit.slice(0, 7)}.`))
+  }
+
+  core.setOutput('version', 'version' in version ? version.version : version.commit)
   core.setOutput('cache-hit', cacheHit)
 }