+++ src/resolve/registry.ts
@@ -3,12 +3,10 @@ import * as core from '@actions/core'
 import { scope } from 'arktype'
 import semver from 'semver'
 
-import { ActionInput } from './parse'
+import type { ResolvedVersion } from '../install'
 
-// Partial response model from crates.io API
-type CrateResponse = typeof types.response.infer
-
-const types = scope({
+// Partial response from crates.io API
+const crateResponseTypes = scope({
   response: {
     crate: {
       max_stable_version: 'semver'
@@ -21,32 +19,36 @@ const types = scope({
   }
 }).compile()
 
+type CrateResponse = typeof crateResponseTypes.response.infer
+
 // Resolve latest compatible crate version
-export async function resolveVersion (input: ActionInput): Promise<string> {
-  const res = await fetchCrate(input.crate)
+export async function resolveRegistryVersion (crate: string, version: string): Promise<ResolvedVersion> {
+  core.info(`Fetching information for ${crate} on crates.io ...`)
+  const res = await fetchCrate(crate)
 
   const latest = res.crate.max_stable_version
-  if (input.version === 'latest') {
-    return latest
+  if (version === 'latest') {
+    return { version: latest }
   }
 
   const resolved = res.versions
-    .filter(ver => semver.satisfies(ver.num, input.version))
+    .filter(ver => semver.satisfies(ver.num, version))
     .sort((a, b) => semver.compare(a.num, b.num))
     .reverse()
   if (resolved.length === 0) {
-    core.setFailed(`No version found for ${input.crate} that satisfies ${input.version}`)
+    core.setFailed(`No version found for ${crate} that satisfies ${version}`)
+    core.info(`Available versions: ${res.versions.map(ver => ver.num).join(', ')}`)
     process.exit(1)
   }
 
-  const version = resolved.find(ver => !ver.yanked) ?? resolved[0]
-  if (version.yanked) {
-    core.warning(`Using yanked version ${version.num} for ${input.crate}`)
-  } else if (version.num !== latest) {
-    core.warning(`New version for ${input.crate} available: ${latest}`)
+  const resolvedVersion = resolved.find(ver => !ver.yanked) ?? resolved[0]
+  if (resolvedVersion.yanked) {
+    core.warning(`Using yanked version ${resolvedVersion.num} for ${crate}`)
+  } else if (resolvedVersion.num !== latest) {
+    core.warning(`New version for ${crate} available: ${latest}`)
   }
 
-  return version.num
+  return { version: resolvedVersion.num }
 }
 
 async function fetchCrate (name: string): Promise<CrateResponse> {
@@ -62,7 +64,7 @@ async function fetchCrate (name: string): Promise<CrateResponse> {
     process.exit(1)
   }
 
-  const { data, problems } = types.response(response.result)
+  const { data, problems } = crateResponseTypes.response(response.result)
 
   if (data === undefined) {
     core.setFailed(`Failed to parse crates.io API response for ${name}`)