+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Git installation is now supported with the `git` input parameter. You can
+  specify a branch, tag or commit hash.
 
 ## [2.0.0] - 2023-03-23
 ### Added